+++ exporters/otlp/src/otlp_grpc_exporter_options.cc
@@ -31,6 +31,8 @@ OtlpGrpcExporterOptions::OtlpGrpcExporterOptions()
   user_agent = GetOtlpDefaultUserAgent();
 
   max_threads = 0;
+
+  compression = GetOtlpDefaultTracesCompression();
 #ifdef ENABLE_ASYNC_EXPORT
   max_concurrent_requests = 64;
 #endif