+++ exporters/zipkin/test/zipkin_exporter_test.cc
@@ -65,14 +65,16 @@ public:
               (const nostd::string_view &,
                const ext::http::client::HttpSslOptions &,
                const ext::http::client::Body &,
-               const ext::http::client::Headers &),
+               const ext::http::client::Headers &,
+               const ext::http::client::Compression &),
               (noexcept, override));
 
   MOCK_METHOD(ext::http::client::Result,
               Get,
               (const nostd::string_view &,
                const ext::http::client::HttpSslOptions &,
-               const ext::http::client::Headers &),
+               const ext::http::client::Headers &,
+               const ext::http::client::Compression &),
               (noexcept, override));
 };
 
@@ -153,7 +155,7 @@ TEST_F(ZipkinExporterTestPeer, ExportJsonIntegrationTest)
 
   auto expected_url = nostd::string_view{"http://localhost:9411/api/v2/spans"};
 
-  EXPECT_CALL(*mock_http_client, Post(expected_url, _, IsValidMessage(report_trace_id), _))
+  EXPECT_CALL(*mock_http_client, Post(expected_url, _, IsValidMessage(report_trace_id), _, _))
 
       .Times(Exactly(1))
       .WillOnce(Return(ByMove(ext::http::client::Result{
@@ -179,7 +181,7 @@ TEST_F(ZipkinExporterTestPeer, ShutdownTest)
   // exporter should not be shutdown by default
   nostd::span<std::unique_ptr<sdk::trace::Recordable>> batch_1(&recordable_1, 1);
 
-  EXPECT_CALL(*mock_http_client, Post(_, _, _, _))
+  EXPECT_CALL(*mock_http_client, Post(_, _, _, _, _))
 
       .Times(Exactly(1))
       .WillOnce(Return(ByMove(ext::http::client::Result{