+++ test_common/include/opentelemetry/test_common/ext/http/client/nosend/http_client_nosend.h
@@ -61,6 +61,12 @@ public:
     timeout_ms_ = timeout_ms;
   }
 
+  void SetCompression(
+      const opentelemetry::ext::http::client::Compression &compression) noexcept override
+  {
+    compression_ = compression;
+  }
+
 public:
   opentelemetry::ext::http::client::Method method_;
   opentelemetry::ext::http::client::HttpSslOptions ssl_options_;
@@ -68,6 +74,8 @@ public:
   opentelemetry::ext::http::client::Headers headers_;
   std::string uri_;
   std::chrono::milliseconds timeout_ms_{5000};  // ms
+  opentelemetry::ext::http::client::Compression compression_{
+      opentelemetry::ext::http::client::Compression::kNone};
 };
 
 class Response : public opentelemetry::ext::http::client::Response