+++ exporters/otlp/include/opentelemetry/exporters/otlp/otlp_environment.h
@@ -41,6 +41,10 @@ std::string GetOtlpDefaultHttpTracesEndpoint();
 std::string GetOtlpDefaultHttpMetricsEndpoint();
 std::string GetOtlpDefaultHttpLogsEndpoint();
 
+std::string GetOtlpDefaultHttpTracesProtocol();
+std::string GetOtlpDefaultHttpMetricsProtocol();
+std::string GetOtlpDefaultHttpLogsProtocol();
+
 // Compatibility with OTELCPP 1.8.2
 inline std::string GetOtlpDefaultHttpEndpoint()
 {
@@ -147,6 +151,10 @@ inline OtlpHeaders GetOtlpDefaultHeaders()
 {
   return GetOtlpDefaultTracesHeaders();
 }
+
+std::string GetOtlpDefaultTracesCompression();
+std::string GetOtlpDefaultMetricsCompression();
+std::string GetOtlpDefaultLogsCompression();
 
 }  // namespace otlp
 }  // namespace exporter