+++ sdk/test/metrics/attributes_hashmap_benchmark.cc
@@ -36,7 +36,7 @@ void BM_AttributseHashMap(benchmark::State &state)
     m.lock();
     auto hash = opentelemetry::sdk::common::GetHashForAttributeMap(attributes[i % 2]);
     hash_map.GetOrSetDefault(attributes[i % 2], create_default_aggregation, hash)
-        ->Aggregate((int64_t)1);
+        ->Aggregate(static_cast<int64_t>(1));
     benchmark::DoNotOptimize(hash_map.Has(hash));
     m.unlock();
   };